separate resolvers into submodule
