+++ hima/experiments/temporal_pooling/new/resolvers/dataset_resolver.py
@@ -3,8 +3,9 @@
 #  All rights reserved.
 #
 #  Licensed under the AGPLv3 license. See LICENSE in the project root for license information.
+
 from hima.common.config_utils import extracted_type, resolve_init_params, TConfig
-from hima.experiments.temporal_pooling.new.blocks.graph_resolver import BlockResolver
+from hima.experiments.temporal_pooling.new.resolvers.graph_resolver import BlockResolver
 from hima.experiments.temporal_pooling.new.blocks.graph import Block
