+++ hima/experiments/temporal_pooling/new/resolvers/encoder_resolver.py
