+++ hima/experiments/temporal_pooling/new/resolvers/sp_resolver.py
@@ -3,9 +3,9 @@
 #  All rights reserved.
 #
 #  Licensed under the AGPLv3 license. See LICENSE in the project root for license information.
-from hima.common.config_utils import TConfig, extracted_type, resolve_init_params
+from hima.common.config_utils import TConfig, resolve_init_params
 from hima.experiments.temporal_pooling.new.blocks.graph import Block
-from hima.experiments.temporal_pooling.new.blocks.graph_resolver import BlockResolver
+from hima.experiments.temporal_pooling.new.resolvers.graph_resolver import BlockResolver
 
 
 class SpatialPoolerResolver(BlockResolver):
